move EngineFunc[Span][Iter] into its own submodule
