+++ crates/wasmi/src/engine/code_map/mod.rs
@@ -5,25 +5,26 @@
 //! This is the data structure specialized to handle compiled
 //! register machine based bytecode functions.
 
+mod span;
+
+pub use self::span::{EngineFunc, EngineFuncSpan, EngineFuncSpanIter};
 use super::{FuncTranslationDriver, FuncTranslator, TranslationError, ValidatingFuncTranslator};
 use crate::{
     Config,
     Error,
     TrapCode,
-    collections::arena::{Arena, ArenaKey},
+    collections::arena::Arena,
     core::{Fuel, FuelCostsProvider},
     engine::{ResumableOutOfFuelError, utils::unreachable_unchecked},
     errors::FuelError,
-    ir::index::InternalFunc,
     module::{FuncIdx, ModuleHeader},
 };
 use alloc::boxed::Box;
 use core::{
     fmt,
     mem::{self, MaybeUninit},
-    ops::{self, Range},
+    ops::Range,
     pin::Pin,
-    slice,
 };
 use spin::Mutex;
 use wasmparser::{FuncToValidate, ValidatorResources, WasmFeatures};
@@ -31,183 +32,11 @@ use wasmparser::{FuncToValidate, ValidatorResources, WasmFeatures};
 #[cfg(doc)]
 use crate::ir::Op;
 
-/// A reference to a compiled function stored in the [`CodeMap`] of an [`Engine`](crate::Engine).
-#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
-pub struct EngineFunc(u32);
-
-impl From<EngineFunc> for InternalFunc {
-    fn from(value: EngineFunc) -> Self {
-        InternalFunc::from(value.0)
-    }
-}
-
-impl From<InternalFunc> for EngineFunc {
-    fn from(index: InternalFunc) -> Self {
-        Self(u32::from(index))
-    }
-}
-
-impl EngineFunc {
-    /// Creates a new [`EngineFunc`] from the given `u32` index.
-    ///
-    /// # Note
-    ///
-    /// This is a test-only API and not meant for code outside of tests.
-    #[cfg(test)]
-    pub fn from_u32(index: u32) -> Self {
-        Self(index)
-    }
-}
-
-impl ArenaKey for EngineFunc {
-    fn into_usize(self) -> usize {
-        self.0.into_usize()
-    }
-
-    fn from_usize(value: usize) -> Option<Self> {
-        <_ as ArenaKey>::from_usize(value).map(Self)
-    }
-}
-
 /// Datastructure to efficiently store information about compiled functions.
 #[derive(Debug)]
 pub struct CodeMap {
     funcs: Mutex<Arena<EngineFunc, FuncEntity>>,
     features: WasmFeatures,
-}
-
-/// A range of [`EngineFunc`]s with contiguous indices.
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct EngineFuncSpan {
-    start: EngineFunc,
-    end: EngineFunc,
-}
-
-impl Default for EngineFuncSpan {
-    #[inline]
-    fn default() -> Self {
-        Self::empty()
-    }
-}
-
-impl EngineFuncSpan {
-    /// Creates a new [`EngineFuncSpan`] for `start..end`.
-    ///
-    /// # Panics
-    ///
-    /// If `start` index is not less than or equal to `end` index.
-    pub fn new(start: EngineFunc, end: EngineFunc) -> Self {
-        assert!(start <= end);
-        Self { start, end }
-    }
-
-    /// Creates an empty [`EngineFuncSpan`].
-    #[inline]
-    pub fn empty() -> Self {
-        Self {
-            start: EngineFunc(0),
-            end: EngineFunc(0),
-        }
-    }
-
-    /// Returns `true` if `self` is empty.
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        debug_assert!(self.start <= self.end);
-        self.start == self.end
-    }
-
-    /// Returns the number of [`EngineFunc`] in `self`.
-    pub fn len(&self) -> u32 {
-        debug_assert!(self.start <= self.end);
-        let start = self.start.0;
-        let end = self.end.0;
-        end - start
-    }
-
-    /// Returns the n-th [`EngineFunc`] in `self`, if any.
-    ///
-    /// Returns `None` if `n` is out of bounds.
-    pub fn get(&self, n: u32) -> Option<EngineFunc> {
-        debug_assert!(self.start <= self.end);
-        if n >= self.len() {
-            return None;
-        }
-        Some(EngineFunc(self.start.0 + n))
-    }
-
-    /// Returns the `u32` index of the [`EngineFunc`] in `self` if any.
-    ///
-    /// Returns `None` if `func` is not contained in `self`.
-    pub fn position(&self, func: EngineFunc) -> Option<u32> {
-        debug_assert!(self.start <= self.end);
-        if func < self.start || func >= self.end {
-            return None;
-        }
-        Some(func.0 - self.start.0)
-    }
-
-    /// Returns the n-th [`EngineFunc`] in `self`, if any.
-    ///
-    /// # Panics
-    ///
-    /// If `n` is out of bounds.
-    #[track_caller]
-    pub fn get_or_panic(&self, n: u32) -> EngineFunc {
-        debug_assert!(self.start <= self.end);
-        self.get(n)
-            .unwrap_or_else(|| panic!("out of bounds `EngineFunc` index: {n}"))
-    }
-
-    /// Returns an iterator over the [`EngineFunc`]s in `self`.
-    #[inline]
-    pub fn iter(&self) -> EngineFuncSpanIter {
-        debug_assert!(self.start <= self.end);
-        EngineFuncSpanIter { span: *self }
-    }
-}
-
-#[derive(Debug)]
-pub struct EngineFuncSpanIter {
-    span: EngineFuncSpan,
-}
-
-impl Iterator for EngineFuncSpanIter {
-    type Item = EngineFunc;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.span.is_empty() {
-            return None;
-        }
-        let func = self.span.start;
-        self.span.start = EngineFunc(self.span.start.0 + 1);
-        Some(func)
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        let remaining = self.span.len() as usize;
-        (remaining, Some(remaining))
-    }
-}
-
-impl DoubleEndedIterator for EngineFuncSpanIter {
-    #[inline]
-    fn next_back(&mut self) -> Option<Self::Item> {
-        if self.span.is_empty() {
-            return None;
-        }
-        self.span.end = EngineFunc(self.span.end.0 - 1);
-        Some(self.span.end)
-    }
-}
-
-impl ExactSizeIterator for EngineFuncSpanIter {
-    #[inline]
-    fn len(&self) -> usize {
-        self.span.len() as usize
-    }
 }
 
 impl CodeMap {
